chore: move context requiring code to a different module
